Add language selector to bottom and move login to navbar
