+++ app/javascript/components/App.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import { Deliveries, Home, Groceries, MasterNavbar, Order, Users } from "../components";
+import { Deliveries, Home, Groceries, LanguageSelector, MasterNavbar, Order, Users } from "../components";
 
 const App = (): JSX.Element => (
     <>
-        <MasterNavbar />
         <Router>
+            <MasterNavbar />
             <Switch>
                 <Route path="/" exact component={Home} />
                 <Route path="/groceries" exact component={Groceries} />
@@ -14,6 +14,7 @@ const App = (): JSX.Element => (
                 <Route path="/users" exact component={Users} />
             </Switch>
         </Router>
+        <LanguageSelector />
     </>
 )
