+++ app/javascript/components/Home/Home.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { TFunction } from 'i18next';
-import { Button, CardDeck, Col, Container, Row, Image } from 'react-bootstrap';
-import { Link } from "react-router-dom";
+import { CardDeck, Col, Container, Row, Image } from 'react-bootstrap';
 import LongLogo from '../../../assets/images/long_logo.png';
 import { HomeCard } from './components';
 
@@ -54,27 +53,6 @@ function getDeliverCard(t: TFunction): JSX.Element {
     );
 }
 
-function handleLogOut(): void {
-    const logOutUrl = "/users/sign_out";
-    const token = document.querySelector('meta[name="csrf-token"]')?.getAttribute("content") || "";
-
-    const params = {
-        method: "DELETE",
-        headers: {
-            "X-CSRF-Token": token,
-            "Content-Type": "application/json"
-        },
-    }
-
-    fetch(logOutUrl, params).then((response: Response) => {
-        if (response.ok) {
-            window.location.reload();
-            return response.json();
-        }
-        throw new Error("Network response was not ok on user sign out.");
-    })
-}
-
 function Home(): JSX.Element {
     const { t } = useTranslation();
 
@@ -90,10 +68,6 @@ function Home(): JSX.Element {
                     </CardDeck>
                 </Col>
             </Row>
-            <Link to='/users'>
-                Log in
-            </Link>
-            <Button onClick={handleLogOut}>Log out</Button>
         </Container>
     );
 }
