+++ app/javascript/components/LanguageSelector/index.ts
@@ -1,0 +1,1 @@
+export { default as LanguageSelector } from './LanguageSelector';
