+++ app/javascript/components/MasterNavbar/MasterNavbar.tsx
@@ -1,10 +1,31 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
+import { Link } from "react-router-dom";
 import { Button, Form, Navbar, Nav } from 'react-bootstrap';
-import { TFunction } from 'i18next';
+
+function handleLogOut(): void {
+    const logOutUrl = "/users/sign_out";
+    const token = document.querySelector('meta[name="csrf-token"]')?.getAttribute("content") || "";
+
+    const params = {
+        method: "DELETE",
+        headers: {
+            "X-CSRF-Token": token,
+            "Content-Type": "application/json"
+        },
+    }
+
+    fetch(logOutUrl, params).then((response: Response) => {
+        if (response.ok) {
+            window.location.reload();
+            return response.json();
+        }
+        throw new Error("Network response was not ok on user sign out.");
+    })
+}
 
 function MasterNavbar(): JSX.Element {
-    const { t, i18n } = useTranslation();
+    const { t } = useTranslation();
 
     return (
         <Navbar bg="light" variant="light">
@@ -16,8 +37,10 @@ function MasterNavbar(): JSX.Element {
                 <Nav.Link href="#about">{t('master_navbar.about')}</Nav.Link>
             </Nav>
             <Form inline>
-                <Button className="mr-1" onClick={(): Promise<TFunction> => i18n.changeLanguage('en')}>English</Button>
-                <Button onClick={(): Promise<TFunction> => i18n.changeLanguage('fr')}>Français</Button>
+                <Link to='/users'>
+                    <Button className="mr-1">Log in</Button>
+                </Link>
+                <Button onClick={handleLogOut}>Log out</Button>
             </Form>
         </Navbar >
     );
