+++ app/javascript/components/index.ts
@@ -4,3 +4,4 @@ export { Order } from './Order';
 export { MasterNavbar } from './MasterNavbar';
 export { Deliveries } from './Deliveries';
 export { Users } from './Users';
+export { LanguageSelector } from './LanguageSelector';
